+++ lib/utils/Lazy.ts
@@ -21,11 +21,10 @@ export class Lazy <T, Env extends Record<string, any> = Record<string, any>> {
 
     resolve (opts:{env:Env}):T {
         if (this.#val) return this.#val;
-
+        
         if (!this.#fn) throw new Error('Lazy@resolve: No initializer provided');
-
+        
         this.#val = this.#fn(opts);
-
         return this.#val;
     }