+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0] - 2025-05-10
 ### Improved
 - **feat**: `ctx.setHeader` now accepts `number` values and will coerce them to strings internally.
 - **feat**: `ctx.setHeaders` now accepts `number` values in the object and coerces them to strings.